requirements.txt
tqdm==4.19.5
Werkzeug==0.12.2
opencv_python==3.3.1.11
numpy==1.13.3
Flask==0.12.2
tensorflow_gpu==0.12.1
Keras==2.1.1
Pillow==4.3.0
tensorflow==1.4.1
